server/api/views/history.py
from rest_framework.decorators import api_view
import json
import hashlib
import datetime
from dateutil.parser import parse
from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
from api.apps import *
import api.auth
from lib.TGMT.TGMTpaging import Paging

####################################################################################################
