requirements.txt
djangorestframework
mongoengine==0.29.1
pymongo==4.9.1
python-dateutil
Pyjwt
Django==2.2.4
django-cors-headers
nose
Sphinx==1.5.5
sphinx-rtd-theme==0.2.4
openpyxl==2.5.6
django-crontab==0.7.1
Pillow==10.1.0
Unidecode==1.3.8
